MOve String utility in opm-common
